add first UI

+++ app/src/main/java/ru/divizdev/creditcalculator/BL/Calculator.java
@@ -4,7 +4,7 @@ package ru.divizdev.creditcalculator.BL;
  * Created by diviz on 22.07.2017.
  */
 public class Calculator {
-    public ICalculation calculation(int months, int interestRate, int amountCredit) {
+    public ICalculation calculation(int months, double interestRate, int amountCredit) {
         OptionsCredit optionsCredit = new OptionsCredit(months, interestRate, amountCredit);
         return new AnnuityCalculation(optionsCredit);
     }
